Book factory

+++ src/model/MemberAccountFactory.java
@@ -1,17 +1,17 @@
 package model;
 
 public class MemberAccountFactory {
-	
-	public static MemberAccountImpl createMemberAndAccount(String id, String email, String firstName, int isSuperMember, String lastName, String user, String pwd) {
+
+	public static MemberAccountImpl createMemberAndAccount(String id, String email, String firstName,
+			byte isSuperMember, String lastName, String user, String pwd) {
 		Member member = new Member(id, email, firstName, isSuperMember, lastName);
 		Account account = new Account(user, pwd);
-		
+
 		member.setAccount(account);
 		account.setMember(member);
-		
-		return new  MemberAccountImpl(member, account);
+
+		return new MemberAccountImpl(member, account);
 	}
-	
 	
 	public static MemberAccountImpl signUp(String user, String pwd) {
 		Member member = new Member();
